+++ cf/regrid/regrid.py
@@ -1132,11 +1132,12 @@ def spherical_grid(
             The grid definition.
 
     """
+    # Create any implied lat/lon coordinates in-place
     try:
-        # Try to convert a HEALPix grid to UGRID
+        # Try to convert a HEALPix grid to a UGRID grid (which will
+        # create 1-d lat/lon coordinates)
         f.healpix_to_ugrid(inplace=True)
     except ValueError:
-        # Create any implied lat/lon coordinates
         f.create_latlon_coordinates(inplace=True)
 
     data_axes = f.constructs.data_axes()